Add Window klc source file and installer package
Included the installer files because not everyone has, or wants, the Microsoft Keyboard Layout Manager program.

Personally I would grab the klc source and build it directly on my own machine but at least you have options this way.

Also, to not lose history this commit is the initial 1.0 version of the klc.  I'll be adding the others after an initial commit.